Extension of ChatGPT only plugin to support private ai services.
 - Initial changes to this project will include renaming of the plugin
to ai-code-review, along with its assets.
 - The new plugin version resides in plugins/ai-code-review and was
 performed to allow Open Source contributions, without impacting chatGPT specific development.
 - Items that overlap will be renamed from chatGPTXXX to chatAIXXX.
Configuration will be renamed where it can be from gptModel / gptToken to aiToken etc. as it is applicable to more than just chatGPT.
 - Creation of new aiType member to allow the switching between different
authorization headers, and API endpoints easily.

- Adding plugin-repo-mvn jenkins build file, so it should build
 master for the submitted changes, and a verifier branch for code reviews.

Testing:
aiType=ollama
aiModel=qwen2.5-coder

OLLAMA working with several models, current suggestion is
to use the qwen-coder model ( note use the 7b model as the
1.5b currently doesn't correctly use tools support ).
The current prereq for the models is that
they support "tools" so that the format_replies tool call
can be made which returns the replies object in JSON matching
the supplied schema, without this it would require much more
parsing to change the response markdown->json on the response.

- Added tests covering new configuration elements, and exposed some
 of the keys to the mocking to ensure they use the same altered names.

TODO:
 - Check what changes are required for Stateful OLLAMA reviews.
 - AzureOpenAI has not been tested - WIP.
   - Auth header coded - using api-key.
   - API item called: apiversion needs to be passed as part of the endpoint string currently but should be a seperate configuration option and added as a query param.

Change-Id: Ic3ddeff429fcb4786eea1ed346eafd76dd6d0976