Extension of ChatGPT only plugin to support private solutions.
Initial changes to this project will include renaming of the plugin
to ai-code-review, along with its assets.
Items that overlap will be renamed from chatGPTXXX to chatAIXXX.
Configuration will be renamed where it can be from gptModel / gptToken to aiToken etc. as it is applicable to more than just chatGPT.

Creation of new aiType member to allow the switching between different
authorization headers, and API endpoints easily.

TODO: - Update readme to match in all cases ( in - progress )
 - Rename / refactor classes to abstract away from chatGPT, as most
 of the model and api members overlap and are general AI Chat apis.
Change-Id: Ic3ddeff429fcb4786eea1ed346eafd76dd6d0976