Extension of ChatGPT only plugin to support private solutions.
- Initial changes to this project will include renaming of the plugin
to ai-code-review, along with its assets.
- Items that overlap will be renamed from chatGPTXXX to chatAIXXX.
- Configuration will be renamed where it can be from gptModel / gptToken to aiToken etc. as it is applicable to more than just chatGPT.
- Creation of new aiType member to allow the switching between different authorization headers, and API endpoints easily.

OLLAMA working with several models, current suggestion is
to use the qwen-coder model ( note use the 7b model as the
1.5b currently doesn't correctly use tools support ).
The current prereq for the models is that
they support "tools" so that the format_replies tool call
can be made which returns the replies object in JSON matching
the supplied schema, without this it would require much more
parsing to change the response markdown->json on the response.

TODO: - Update readme to match in all cases ( in - progress )
 - Rename / refactor classes to abstract away from chatGPT, as most
 of the model and api members overlap and are general AI Chat apis.
Change-Id: Ic3ddeff429fcb4786eea1ed346eafd76dd6d0976